requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.31
botocore==1.17.31
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.4
cs50==5.0.4
cycler==0.11.0
Cython==0.29.32
distlib==0.3.1
Django==3.1.4
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
django-matplotlib==0.1
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
flatbuffers==2.0.7
fonttools==4.37.1
gluonnlp==0.10.0
graphviz==0.8.4
gunicorn==20.0.4
huggingface-hub==0.0.12
idna==2.9
install==1.3.3
isort==5.6.4
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
JPype1==1.4.0
kiwisolver==1.4.4
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3
konlpy==0.6.0
lazy-object-proxy==1.4.3
lxml==4.9.1
MarkupSafe==1.1.1
matplotlib==3.5.3
mccabe==0.6.1
mxnet==1.7.0.post2
nltk==3.7
numpy==1.23.2
onnxruntime==1.8.0
packaging==21.3
pandas==1.4.4
Pillow==8.0.1
protobuf==4.21.5
psycopg2==2.9.3
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==6.0
regex==2022.8.17
requests==2.23.0
s3transfer==0.3.3
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.1
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.10.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.8.1
typing_extensions==4.3.0
urllib3==1.25.9
virtualenv==20.2.1
Werkzeug==0.15.4
whitenoise==5.1.0
wordcloud==1.8.2.2
wrapt==1.12.1
